tuning and separation of MSSPI_CERT
